documentation, input validation, offline validation and ui fixes

+++ PhoneNumberChecker/Models/NumberValidationPayload.cs
@@ -4,7 +4,7 @@ namespace PhoneNumberChecker.Models
     public partial class NumberValidationPayload
     {
         public int countryCode { get; set; }
-        public int phoneNumber { get; set; }
+        public ulong phoneNumber { get; set; }
 
         public string formattedNbr => phoneNumber.ToString();
     }
